dom/html/test/test_bug613113.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=613113
-->
<head>
  <title>Test for Bug 613113</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=613113">Mozilla Bug 613113</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe name='f'></iframe>
  <form target='f' action="data:text/html,">
    <output></output>
    <button></button>
  </form>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 613113 **/

SimpleTest.waitForExplicitFinish();

var invalidEvent = false;

var form = document.forms[0];
var button = document.getElementsByTagName('button')[0];
var output = document.getElementsByTagName('output')[0];

output.addEventListener("invalid", function() {
  ok(false, "invalid event should have been send");
});

form.addEventListener("submit", function() {
  ok(true, "submit has been caught");
  setTimeout(function() {
    SimpleTest.finish();
  }, 0);
});

output.setCustomValidity("foo");

button.click();

</script>
</pre>
</body>
</html>
